# docs/swagger.yaml
basePath: /
definitions:
...
  dto.FineVoidRequest:
    properties:
      notes:
        type: string
    type: object
...
  utils.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      meta: {}
      status:
        type: boolean
    type: object
host: localhost:8888
info:
  contact:
    email: mddx56@gmail.com
    name: API Support
    url: http://www.swagger.io/support
  description: API para gestionar motos
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: RutaSegura API
  version: 0.1.2
paths:
...
  /api/fine-types:
    get:
      consumes:
      - application/json
      description: Get the catalog of fine types (overtaking, lap time, prolonged
        stop, speeding)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List all fine types
      tags:
      - fines
  /api/fines:
    get:
      consumes:
      - application/json
      description: Get a list of all fines generated for micros (automatic or manual)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List all fines (admin)
      tags:
      - fines
  /api/fines/{id}:
    get:
      consumes:
      - application/json
      description: Get a fine by ID, including its fine type
      parameters:
      - description: Fine ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get a fine by ID
      tags:
      - fines
  /api/fines/{id}/void:
    patch:
      consumes:
      - application/json
      description: Mark a fine as voided after human review (admin only)
      parameters:
      - description: Fine ID
        in: path
        name: id
        required: true
        type: string
      - description: Void reason
        in: body
        name: fine
        schema:
          $ref: '#/definitions/dto.FineVoidRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Void a fine
      tags:
      - fines
  /api/fines/mine:
    get:
      consumes:
      - application/json
      description: Get the fines generated for the authenticated user's own vehicles
        (rol Dueño)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List my fines
      tags:
      - fines
...
  /api/laps:
    get:
      consumes:
      - application/json
      description: Get a list of all laps recorded by the lap-counting engine
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List all laps
      tags:
      - laps
  /api/laps/{id}:
    get:
      consumes:
      - application/json
      description: Get a lap by ID
      parameters:
      - description: Lap ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get a lap by ID
      tags:
      - laps
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"